requirements.txt
moviepy>=2.0.0
pillow
numpy
